+++ Project.BLL/ManagerServices/Abstracts/IMenuDetailManager.cs
@@ -10,10 +10,10 @@ namespace Project.BLL.ManagerServices.Abstracts
 {
     public interface IMenuDetailManager : IManager<MenuDetail>
     {
-        Task<IEnumerable<object>> Get_FoodsofMenu_Async(int Menu_ID);
-        Task<IEnumerable<CategoriesOfMenu_Repo>> Get_CategoriesofMenu_Async(int Menu_ID);
-        Task<bool> IsExist_FoodinMenu_Async(int selected_foodID, int menu_ID);
-        void Insert_FoodonMenu_Async(int selected_foodID, string category_Name, int menu_ID);
+        Task<IEnumerable<object>> Get_FoodsofMenu_Async(short Menu_ID);
+        Task<IEnumerable<CategoriesOfMenu_Repo>> Get_CategoriesofMenu_Async(short Menu_ID);
+        Task<bool> IsExist_FoodinMenu_Async(short selected_foodID, short menu_ID);
+        void Insert_FoodonMenu_Async(short selected_foodID, string category_Name, short menu_ID);
 
 
     }
