+++ Project.BLL/ManagerServices/Concretes/MenuDetailManager.cs
@@ -20,7 +20,7 @@ namespace Project.BLL.ManagerServices.Concretes
             _mdrep = mdrep;
         }
 
-        public async Task<IEnumerable<object>> Get_FoodsofMenu_Async(int Menu_ID)
+        public async Task<IEnumerable<object>> Get_FoodsofMenu_Async(short Menu_ID)
         {            
             var Foods = await _mdrep.Get_FoodsofMenu_Async(Menu_ID).ToListAsync(); 
             // convert ıqueryable to IEnumerable (using namespace EntityFrameworkCore)
@@ -28,7 +28,7 @@ namespace Project.BLL.ManagerServices.Concretes
             return Foods;
         }
 
-         public async Task<IEnumerable<CategoriesOfMenu_Repo>> Get_CategoriesofMenu_Async(int Menu_ID)
+         public async Task<IEnumerable<CategoriesOfMenu_Repo>> Get_CategoriesofMenu_Async(short Menu_ID)
         {
 
             var Categories = await _mdrep.Get_CategoriesofMenu_Async(Menu_ID).ToListAsync();
@@ -38,14 +38,14 @@ namespace Project.BLL.ManagerServices.Concretes
             return Categories;
         }
 
-        public async Task<bool> IsExist_FoodinMenu_Async(int selected_foodID, int menu_ID)
+        public async Task<bool> IsExist_FoodinMenu_Async(short selected_foodID, short menu_ID)
         {
             bool food_exists = await _mdrep.IsExist_FoodinMenu_Repo_Async(selected_foodID, menu_ID);//.ToListAsync();
 
             return food_exists;
         }
 
-        public void Insert_FoodonMenu_Async(int selected_foodID, string category_Name, int menu_ID)
+        public void Insert_FoodonMenu_Async(short selected_foodID, string category_Name, short menu_ID)
         {
             _mdrep.Insert_FoodonMenu_Repo_Async(selected_foodID, category_Name, menu_ID);//.ToListAsync();
         }
