+++ Project.DAL/Repositories/Abstracts/IMenuDetailRepository.cs
@@ -10,10 +10,10 @@ namespace Project.DAL.Repositories.Abstracts
 {
     public interface IMenuDetailRepository : IRepository<MenuDetail>
     {
-        IQueryable<object> Get_FoodsofMenu_Async(int Menu_ID);
-        IQueryable<CategoriesOfMenu_Repo> Get_CategoriesofMenu_Async(int Menu_ID);
-        Task<bool> IsExist_FoodinMenu_Repo_Async(int selected_foodID, int menu_ID);
-        void Insert_FoodonMenu_Repo_Async(int selected_foodID, string category_Name, int menu_ID);
+        IQueryable<object> Get_FoodsofMenu_Async(short Menu_ID);
+        IQueryable<CategoriesOfMenu_Repo> Get_CategoriesofMenu_Async(short Menu_ID);
+        Task<bool> IsExist_FoodinMenu_Repo_Async(short selected_foodID, short menu_ID);
+        void Insert_FoodonMenu_Repo_Async(short selected_foodID, string category_Name, short menu_ID);
 
 
     }
