+++ Project.DAL/Repositories/Concretes/MenuDetailRepository.cs
@@ -33,7 +33,7 @@ namespace Project.DAL.Repositories.Concretes
         {
         }
 
-        public IQueryable<object> Get_FoodsofMenu_Async(int Menu_ID)
+        public IQueryable<object> Get_FoodsofMenu_Async(short Menu_ID)
         {
             //object asd = new Deneme();
             /* !!! !!!
@@ -54,7 +54,7 @@ namespace Project.DAL.Repositories.Concretes
             return null;
         }
 
-        public IQueryable<CategoriesOfMenu_Repo> Get_CategoriesofMenu_Async(int Menu_ID)
+        public IQueryable<CategoriesOfMenu_Repo> Get_CategoriesofMenu_Async(short Menu_ID)
        {
            IQueryable<CategoriesOfMenu_Repo> com_list;
 
@@ -74,14 +74,14 @@ namespace Project.DAL.Repositories.Concretes
             return null;
         }
 
-        public async Task<bool> IsExist_FoodinMenu_Repo_Async(int selected_foodID, int menu_ID)
+        public async Task<bool> IsExist_FoodinMenu_Repo_Async(short selected_foodID, short menu_ID)
         {
             bool food_exists = _context.Set<MenuDetail>().Any(x => x.MenuID == menu_ID && x.FoodID == selected_foodID);//.ToListAsync();
 
             return food_exists;
         }
 
-        public void Insert_FoodonMenu_Repo_Async(int selected_foodID, string category_Name, int menu_ID)
+        public void Insert_FoodonMenu_Repo_Async(short selected_foodID, string category_Name, short menu_ID)
         {
 
             MenuDetail menuDetail = new MenuDetail();
